requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
google-generativeai>=0.3.0
pandas>=2.0.0
tqdm>=4.65.0

